refactor: update imports from CategoriesStore to ConfigurationStore for consistency
