+++ app/_layout.tsx
@@ -12,7 +12,7 @@ import 'react-native-reanimated';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { StatusBar } from 'expo-status-bar';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import CategoriesStore from '@/tbStores/CategoriesStore';
+import CategoriesStore from '@/tbStores/ConfigurationStore';
 import { Provider as TinyBaseProvider } from 'tinybase/ui-react';
 
 export {
